BE JWT security done

+++ src/main/java/com/attila/taskmanager/service/TaskService.java
@@ -1,5 +1,6 @@
 package com.attila.taskmanager.service;
 
+import com.attila.taskmanager.domain.AppUser;
 import com.attila.taskmanager.domain.Task;
 import com.attila.taskmanager.dto.request.CreateTaskCommand;
 import com.attila.taskmanager.dto.request.UpdateTaskCommand;
@@ -7,8 +8,10 @@ import com.attila.taskmanager.dto.response.TaskListItem;
 import com.attila.taskmanager.exception.TaskAlreadyExistsException;
 import com.attila.taskmanager.exception.TaskNotFoundException;
 import com.attila.taskmanager.repository.TaskRepository;
+import com.attila.taskmanager.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -22,36 +25,55 @@ public class TaskService {
 
     private final TaskRepository taskRepository;
 
-    public Long createNewTask(CreateTaskCommand createTaskCommand) {
+    private final UserRepository userRepository;
+
+    public Long createNewTask(CreateTaskCommand createTaskCommand, String username) {
         if (taskRepository.existsByName(createTaskCommand.getName())) {
             throw new TaskAlreadyExistsException("Task already exists");
         }
+
+        AppUser appUser = userRepository.findByUsername(username)
+                .orElseThrow(() -> new UsernameNotFoundException("User not found: " + username));
+
         Task task = new Task();
         task.setName(createTaskCommand.getName());
         task.setDescription(createTaskCommand.getDescription());
+        task.setAppUser(appUser);
         log.info("New task created");
         return taskRepository.save(task).getId();
     }
 
     @Transactional(readOnly = true)
-    public List<TaskListItem> getAllTask() {
+    public List<TaskListItem> getAllTask(String username) {
 
-        return taskRepository.findAll()
+        AppUser appUser = userRepository.findByUsername(username)
+                .orElseThrow(()-> new UsernameNotFoundException("User not found: " + username));
+
+        log.info("Get all tasks");
+        return taskRepository.findAllByAppUser(appUser)
                 .stream()
                 .map(task -> new TaskListItem(task.getId(), task.getName(), task.getDescription(), task.isCompleted()))
                 .toList();
     }
 
     @Transactional(readOnly = true)
-    public TaskListItem getItemById(Long id) {
+    public TaskListItem getItemById(Long id, String username) {
 
         Task task = taskRepository.findById(id).orElseThrow(() -> new TaskNotFoundException("Task not found with id: " + id));
+
+        if (!task.getAppUser().getUsername().equals(username)) {
+            throw new TaskNotFoundException("Task not found with id: " + id);
+        }
         return new TaskListItem(task.getId(), task.getName(), task.getDescription(), task.isCompleted());
     }
 
-    public void updateTask(Long id, UpdateTaskCommand updateTaskCommand) {
+    public void updateTask(Long id, UpdateTaskCommand updateTaskCommand, String username) {
 
         Task task = taskRepository.findById(id).orElseThrow(()-> new TaskNotFoundException("Task not found with id: " + id));
+
+        if (!task.getAppUser().getUsername().equals(username)) {
+            throw new TaskNotFoundException("Task not found with id: " + id);
+        }
 
         if (updateTaskCommand.getName() != null) {
             task.setName(updateTaskCommand.getName());
@@ -66,9 +88,13 @@ public class TaskService {
         taskRepository.save(task);
     }
 
-    public void deleteTask(Long id) {
+    public void deleteTask(Long id, String username) {
 
         Task task = taskRepository.findById(id).orElseThrow(()-> new TaskNotFoundException("Task not found with id: " + id));
+
+        if (!task.getAppUser().getUsername().equals(username)) {
+            throw new TaskNotFoundException("Task not found with id: " + id);
+        }
 
         taskRepository.delete(task);
     }
